Break out major functional pieces to modules
